Move AppContextStore to stores dir
