+++ client/src/SignIn.js
@@ -4,7 +4,7 @@ import Input from 'antd/lib/input';
 import message from 'antd/lib/message';
 import React, { useState, useContext, useEffect } from 'react';
 import { Link, Redirect } from 'react-router-dom';
-import AppContext from './containers/AppContext';
+import { AppContext } from './stores/AppContextStore';
 import fetchJson from './utilities/fetch-json.js';
 
 function SignIn(props) {
