--- client/src/containers/AppContext.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const AppContext = React.createContext({ refreshAppContext: () => {} });
-
-export default AppContext;
