+++ client/src/queryEditor/QueryEditorContainer.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import AppNav from '../AppNav';
 import { ConnectionsContext } from '../connections/ConnectionsStore';
-import AppContext from '../containers/AppContext';
+import { AppContext } from '../stores/AppContextStore';
 import QueryEditor from './QueryEditor';
 
 function QueryEditorContainer(props) {
